requirements.txt
aiofiles==0.4.0
aiohttp==3.3.2
aiohttp-session==2.7.0
ansicolors==1.1.8
asyncio==3.4.3
asyncpg==0.18.3
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
colorama==0.4.1
colorlog==4.0.2
colormap==1.0.2
cowsay==2.0.3
cryptography==2.6.1
datadog==0.30.0
dataset==1.1.2
DateTime==4.3
dateutils==0.6.6
dblpy==0.1.6
decorator==4.4.0
discord.py==1.2.3
dnspython==1.16.0
easydev==0.9.37
emoji==0.5.4
futures==3.1.1
gevent==1.4.0
google-api-python-client==1.7.9
google-assistant-grpc==0.2.1
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googleapis-common-protos==1.5.10
greenlet==0.4.15
grequests==0.3.0
grpcio==1.20.1
gTTS==2.0.3
gTTS-token==1.1.3
httplib2==0.12.3
humanfriendly==4.18
humanize==0.5.1
hypixel==0.7.9
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
imgkit==1.0.1
import-expression==0.3.8.post2
isort==4.3.13
jishaku==1.8.0
ksoftapi==0.1.5b0
lazy-object-proxy==1.3.1
lxml==4.3.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
meld3==1.0.2
motor==2.0.0
moviepy==1.0.0
multidict==4.5.2
numpy==1.16.4
oauthlib==3.0.1
opengraph==0.5
opengraph-py3==0.71
pathlib2==2.3.3
pexpect==4.7.0
pillow>=6.2.0
proglog==0.1.9
protobuf==3.7.1
psutil==5.6.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pylint==2.3.1
pymongo==3.8.0
PyNaCl==1.2.1
pyOpenSSL==19.0.0
pyreadline==2.1
python-dateutil==2.8.0
python-editor==1.0.4
pytube==9.4.0
pytz==2018.9
ratelimiter==1.2.0.post0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
sentry-sdk==0.11.2
six==1.12.0
sounddevice==0.3.13
soupsieve==1.9.1
StringGenerator==0.3.3
supervisor==4.0.0.dev0
tenacity==4.12.0
termcolor==1.1.0
tqdm==4.32.1
typed-ast==1.3.1
typing==3.6.6
typing-extensions==3.7.4
ujson==1.35
wavelink==0.2.8
webkit2png==0.8.2
